+++ lib/src/calendar_property.dart
@@ -15,10 +15,13 @@ abstract class CalendarProperty<ValueType extends CalendarValue>
 
   @override
   String toString() {
+    var parameters = getParameters();
+    parameters.sort((a, b) => a.parameterName.compareTo(b.parameterName),);
+
     final res = StringBuffer();
     res.writeAll([
       propertyName.toUpperCase(),
-      ...getParameters(),
+      ...parameters,
       ...value.getInlineParameters()
     ], ";");
     res.write(":");